+++ spec/swagcov/runner_spec.rb
@@ -13,7 +13,9 @@ RSpec.describe Swagcov::Runner do
     it "prints options" do
       expect { init.run }.to raise_exception(SystemExit).and output(
         <<~MESSAGE
-          Usage: swagcov [options]
+          Usage:
+          * as executable: swagcov [options]
+          * as rake task: rake swagcov -- [options]
               -i, --init                       Generate required .swagcov.yml config file
               -t, --todo                       Generate optional .swagcov_todo.yml config file
         MESSAGE